email_verif
In settings.py:

Configure the email backend: In the settings.py file, you need to specify the email backend configuration for your Django project. This includes settings such as EMAIL_BACKEND, EMAIL_HOST, EMAIL_PORT, EMAIL_HOST_USER, EMAIL_HOST_PASSWORD, and EMAIL_USE_TLS. These settings are used to connect to your email service provider's SMTP server and send emails.
In views.py:

get_verification_link function: This function generates the verification link for the email. It takes the request and user objects as parameters and returns the verification link. The link is typically constructed using the user's unique identifier (uid) and a verification token.

verify_email function: This function handles the email verification process. It checks the validity of the token and activates the user if the token is valid. The uidb64 and token parameters are decoded and used to retrieve the user from the database. If the user exists and the token is valid, the is_active field of the user is set to True, indicating that the user's email has been verified.

send_verification_email function: This function sends an email to the user containing the verification link message